% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_constants.R
\name{fg_update_aes}
\alias{fg_update_aes}
\alias{fg_get_aes}
\alias{fg_get_aesstring}
\alias{fg_update_line_colors}
\alias{fg_display_colors}
\alias{fg_print_aes_list}
\alias{fg_reset_to_default_state}
\title{Maintain Colors}
\usage{
fg_get_aes(item = "", n_max = NA_integer_, asdataframe = FALSE)

fg_get_aesstring(item = "", n_max = NA_integer_, toget = "value")

fg_update_aes(indta, aestype = NA_character_, persist = TRUE, replace = FALSE)

fg_update_line_colors(colorlist, replace = FALSE, persist = TRUE)

fg_display_colors(item = "")

fg_print_aes_list(grepstr = "")

fg_reset_to_default_state(reset = "all")
}
\arguments{
\item{item}{(Default: "") A grep string for categories desired.}

\item{n_max}{Maximum number of rows or entries to return.}

\item{asdataframe}{(Default: FALSE) Return dataframe of parameters regardless of type. (See details)}

\item{toget}{Column in the aes \code{data.frame} to paste together as a string.}

\item{indta}{\code{data.table} aesthetic \code{data.fram} with columns as shown in details.}

\item{aestype}{(Default: \code{NA}) character string with type of aesthetic requested.  If not provided in \verb{[fg_Update_aes()]} the
aesthetic associated with \code{item} is inferred from existing data.}

\item{persist}{(Default: TRUE) Keep changes across invocations of the package.}

\item{replace}{(Default: FALSE) Replaces existing dates of interest with new set provided, otherwise replaces/inserts new rows only.}

\item{colorlist}{List with up to 14 new colors just for line (series) coloring}

\item{grepstr}{narrow list of internal aesthetics sets to functions from \code{grepstr}}

\item{reset}{(Default: "all"), options in ("all","colors","doi") to reset to defaults with the package.}
}
\value{
\code{data.frame}, \code{string} or message
}
\description{
\code{fg_get_aes()} gets aethestic  \code{data.frame} for use in graphs.
\code{fg_get_aesstring()} takes a column from the \code{data.frame} retrieved by \code{fg_get_aes()}
\code{fg_print_aes_list()} prints names of aesthetics used internally in FinanceGraph functions.
\code{fg_update_aes()} updates or replaces default aesthestics (e.g. colors, linestyles, etc).
\code{fg_display_colors()} Shows a plot with current colors.
\code{fg_update_line_colors()} replaces line colors only
\code{fg_reset_to_default_state()} resets colors and/or dates of interest
}
\details{
For colors,
New data passed into \code{\link[=fg_update_aes]{fg_update_aes()}} persists across future loads of the package unless \code{persist=FALSE}.
New color datasets must have at least three columns:\tabular{ll}{
   Column \tab Meaning \cr
   \code{category} \tab Arbitrary aestehtic category, e.g. \code{"lines"} for line colors. \cr
   \code{variable} \tab Any string that can be sorted or grepped to map to data. \cr
   \code{type} \tab Aesthetic type, in \code{c("color","colorrange","linetype","symbol","alpha")} \cr
   \code{value} \tab String with value detired (e.g a color) \cr
}


\code{variable} is used to prioritize colors, so (e.g. \code{D01} will be the color of the first series in an input dataset)

If \code{aestype=="colorrange"} then a sequential scale of size \code{n_max} will be returned using details saved from  \code{\link[=fg_update_aes]{fg_update_aes()}}. See \link[scales:pal_brewer]{scales::brewer_pal}
and \href{https://colorbrewer2.org/#type=sequential&scheme=Greens&n=7}{colorbrewer}
}
\examples{
# Data set, String
head(fg_get_aes("lines"),3)
fg_get_aesstring("lines")
#  Gradient colors are stored in a `data.frame` as in a set of "Blue Greens"
fg_get_aes("espath_gp",asdataframe=TRUE)
# To get the actual colors, we need to know how many:
fg_get_aes("espath_gp",n_max=8)
# To add aesthetics or update them, get the original data (or make from scratch) with
head(oldcolors <- fg_get_aes("lines"),3)
# then change as needed.  For example, to make the second line blue, and the 4th line red,
oldcolors[c(2,3),"value"] <- c("blue","red")
fg_update_aes( oldcolors )
head( fg_get_aes("lines"),3)
# to create a new category, make a similar `data.frame`, as in
newcolors <- data.frame(category=rep("mylines",3),variable=c("D01","D02","D03"),
                value=c("red","black","green"))
fg_update_aes( newcolors, aestype="color")
fg_get_aesstring("mylines")
require(scales)
show_col(fg_get_aesstring("mylines"))
fg_reset_to_default_state("aes")


}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}
}
