% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{consumer_sent}
\alias{consumer_sent}
\title{Consumer Sentiment Data}
\format{
\subsection{\code{consumer_sent}}{

A data frame with 120 rows and 3 columns
\describe{
\item{symbol}{FRED identifier}
\item{date}{Date of report}
\item{pricce}{Observation}
}
}
}
\source{
\url{https://fred.stlouisfed.org/}
}
\usage{
consumer_sent
}
\description{
University of Michigan Consumer Sentiment Data
}
\keyword{datasets}
