% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{eqtypx_melt}
\alias{eqtypx_melt}
\title{Equity Prices}
\format{
\subsection{\code{eqtypx_melt}}{

A tibble with 10116 observations and 3 columns
\describe{
\item{date}{Date Of Observation}
\item{variable}{Ticker}
\item{value}{Closing Price}
}
}
}
\source{
\url{https://business-science.github.io/tidyquant/}
}
\usage{
eqtypx_melt
}
\description{
Closing Equity Prices (melted)
}
\keyword{datasets}
