% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extenddtstr}
\alias{extenddtstr}
\title{Extend a generic date string in varioud ways}
\usage{
extenddtstr(
  instr,
  begchg = 0,
  endchg = 0,
  mindt = NULL,
  maxdt = NULL,
  rtn = "",
  rtnstyle = "string"
)
}
\arguments{
\item{instr}{Input generalized date string, \code{data.table} or \code{xts} dataset}

\item{begchg}{(Default: 0) Number of calendar days to extend beginning}

\item{endchg}{(Default: 0) Number of calendar days to extend end}

\item{mindt}{Minimum date to return}

\item{maxdt}{Maximum date to return}

\item{rtn}{string describing what to do: (\code{list},\code{datelist},\code{fromtoday},\code{totoday})}

\item{rtnstyle}{REturn datestring or list}
}
\value{
String or list with new dates
}
\description{
Extend a generic date string in varioud ways
}
\examples{
extenddtstr("-2m::-1m")
extenddtstr("-2m::-1m",begchg=10,endchg=5)

}
