% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_addbreakouts}
\alias{fg_addbreakouts}
\title{Add Breakout events to time series plots}
\usage{
fg_addbreakouts(indta, annotationstyle = "singleasdate", ...)
}
\arguments{
\item{indta}{Time series \code{data.table} with a date as the first column and a value series as the second column.}

\item{annotationstyle}{String in set (\code{singleasdate},\code{singleasvalue},'breakno')}

\item{...}{Parameters passed to \code{\link[ecp:e.divisive]{ecp::e.divisive()}}}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Wrapper around the function \code{\link[ecp:e.divisive]{ecp::e.divisive()}} to create events for \code{\link[=fgts_dygraph]{fgts_dygraph()}}
}
\details{
Event Helpers : fg_addbreakouts
}
\examples{
if (requireNamespace("ecp", quietly = TRUE)) {
dta <- tail(eqtypx[,.(date,QQQ,TLT)],2*260)
fgts_dygraph(dta,event_ds=fg_addbreakouts(dta,min.size=66,R=40),title="With Breakouts")
}
}
