% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_av_earnings}
\alias{fg_av_earnings}
\title{Event Helpers: fg_av_earnings}
\usage{
fg_av_earnings(indt, field = "reportedEPS", ticker_in_label = TRUE)
}
\arguments{
\item{indt}{\code{data.frame} obtained from alphavantage earnings data.}

\item{field}{(Default: \code{reportedEPS}) String in (\code{reportedEPS},\code{estimatedEPS},\code{surprise},\code{surprisePercentage})}

\item{ticker_in_label}{(Default: TRUE) Make label ticker and the earnings}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Created \code{event_ds} from \link[alphavantagepf:av_get_pf]{alphavantagepf::av_get_pf} quarterly earnings data.
}
\examples{
require(alphavantagepf)
earnings = alphavantagepf::av_get_pf("IBM","EARNINGS") |>
        alphavantagepf::av_extract_df("quarterlyEarnings") |>
        fg_av_earnings()
toplot = dplyr::select(eqtypx,date,IBM)
fgts_dygraph(toplot,title="With earnings",dtstartfrac=0.8,event_ds=earnings)

}
