% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_cut_to_events}
\alias{fg_cut_to_events}
\title{Cut a univariate time series into colored event bars.}
\usage{
fg_cut_to_events(
  indta,
  ncutsperside = 4,
  center = 0,
  extend = TRUE,
  invert = FALSE
)
}
\arguments{
\item{indta}{Time series \code{data.table} with a date as the first column and a value series as the second column.}

\item{ncutsperside}{: Integer with number of colors to use on each side of 'center'}

\item{center}{: String or Double as follows:
\itemize{
\item Double (default 0) Normalize data by subtracting \code{center}
\item \code{"median"} Normalize data by subtracting median of all observations.
\item \code{"zscore"} Normalize data by using standard \code{\link[=scale]{scale()}} function
}}

\item{extend}{Logical (Default: TRUE) to extend data to today (\code{Sys.Date()})
data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAbElEQVR4Xs2RQQrAMAgEfZgf7W9LAguybljJpR3wEse5JOL3ZObDb4x1loDhHbBOFU6i2Ddnw2KNiXcdAXygJlwE8OFVBHDgKrLgSInN4WMe9iXiqIVsTMjH7z/GhNTEibOxQswcYIWYOR/zAjBJfiXh3jZ6AAAAAElFTkSuQmCC}

\item{invert}{Use opposite color schemes for data, i.e. "red" for good outcomes}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Event Helpers: fg_cut_to_events
}
\details{
Always uses first date column and first numeric columns in data.  If \code{indta} has multiple series, filter them before calling the function.
}
\examples{
smalldta <- narrowbydtstr(eqtypx[,.(date,IBM,QQQ)],"-2y::")
events <- fg_cut_to_events(consumer_sent,center="zscore")
fgts_dygraph(smalldta,title="With Sentiment ranges",event_ds=events)
}
