% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gendtstr}
\alias{gendtstr}
\alias{narrowbydtstr}
\alias{extenddtstr}
\title{Date Utilities}
\usage{
gendtstr(x, today = Sys.Date(), rtn = "dtstr")

narrowbydtstr(
  xin,
  dtstr = "",
  includetoday = TRUE,
  windowdays = 0,
  invert = FALSE,
  addindicator = FALSE
)

extenddtstr(
  instr,
  begchg = 0,
  endchg = 0,
  mindt = NULL,
  maxdt = NULL,
  rtn = "",
  rtnstyle = "string"
)
}
\arguments{
\item{x}{String describing generalized date as of today}

\item{today}{Default \code{Sys.Date()}}

\item{rtn}{string describing what to do: (\code{list},\code{datelist},\code{fromtoday},\code{totoday})}

\item{xin}{Input \code{data.frame} or \code{data.table} with a Date column}

\item{dtstr}{Generalized Date string of the form \verb{<yyyy-mm-dd>::<yyyy-mm-dd>} or e.g. \verb{-3m::}}

\item{includetoday}{(Default: TRUE) pass either today \code{Sys.Date()} or \code{Sys.Date()-1} to \code{gendtstr}}

\item{windowdays}{(Default: 0)Number of additional days to add at beginning of series}

\item{invert}{(Default: FALSE) Return dates not in \code{dtstr}}

\item{addindicator}{(Default: FALSE) Returns original dataset with logical variable \code{inrange} if date is in desired range.}

\item{instr}{Input generalized date string, \code{data.table} or \code{xts} dataset}

\item{begchg}{(Default: 0) Number of calendar days to extend beginning}

\item{endchg}{(Default: 0) Number of calendar days to extend end}

\item{mindt}{Minimum date to return}

\item{maxdt}{Maximum date to return}

\item{rtnstyle}{REturn datestring or list}
}
\value{
an exact start date \code{startdt} and an exact end date \code{enddt}, in the following forms:
If \code{rtn="list"} returns \code{c(startdt,enddt)}, if \code{rtn="first"} then \code{startdt}, if \verb{rtn="days} then an integer number of days from \code{startdt} to \code{today}  otherwise (by default) \code{"startdt::enddt"}

Same form as \code{xin}, i.e. a \code{data.table} or \code{data.frame}

\code{character} string or \code{list} with new dates
}
\description{
COnverts a generic relative string defining one or two endpoints to exact dates or datestrings
}
\examples{
gendtstr("-3m::")
gendtstr("-2y::-3m",today=as.Date("2025-03-15"))
narrowbydtstr(eqtypx,"-2m::-1m")
extenddtstr("-2m::-1m")
extenddtstr("-2m::-1m",begchg=-10,endchg=5)
}
