% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_constants.R
\docType{data}
\name{fg_dates_of_interest}
\alias{fg_dates_of_interest}
\alias{the}
\alias{fg_get_dates_of_interest}
\alias{fg_get_colors}
\alias{fg_update_dates_of_interest}
\alias{fg_update_colors}
\alias{fg_reset_to_default_state}
\title{Maintain Constants and Dates of Interest}
\format{
An object of class \code{environment} of length 10.
}
\usage{
the

fg_get_dates_of_interest(
  search_categories = "",
  use_default = TRUE,
  startdt = NULL
)

fg_get_colors(item = "", n_max = NA_integer_)

fg_update_dates_of_interest(indta, replace = FALSE)

fg_update_colors(indta, replace = FALSE)

fg_reset_to_default_state(reset = "all")
}
\arguments{
\item{search_categories}{Grep string of categories to return.}

\item{use_default}{(Default: TRUE) Use package level dates of interest.}

\item{startdt}{MInimum date for events to be returned.}

\item{item}{(Default: "") A grep string for categories desired.}

\item{n_max}{Maximum number of rows to return.}

\item{indta}{\code{data.table} with columns as shown in details.}

\item{replace}{(Default: FALSE) If TRUE, replaces existing dates of interest with new set provided, otherwise replaces/inserts new rows only.}

\item{reset}{(Default: "all"), options in ("all","colors","doi") to reset to defaults with the package.}
}
\value{
Filtered datasets
}
\description{
\code{fg_get_dates_of_interest()} gets a set of time events for use in fg time series graphs
\code{fg_update_dates_of_interest()} updates a set of time events for future use in time series graphs
\code{fg_get_colors()} gets default color sets for graphs
\code{fg_update_colors()} updates or replaces default colors
\code{fg_reset_to_default_state()} resets colors and/or dates of interest
}
\details{
Retrieves default dates of interest given a grepstring of categories.  There are a default set of categories provided which may not be up to date.

New data passed into \code{\link[=fg_update_dates_of_interest]{fg_update_dates_of_interest()}} or \code{\link[=fg_update_colors]{fg_update_colors()}} persists across future loads of the package. Any duplicates in the new file will be taken out.

New doi datasets must have at least three columns:\tabular{ll}{
   Column \tab Meaning \cr
   \code{category} \tab Grouping name (string) for a given set of dates of interest \cr
   \code{eventid} \tab Character string to be displayed at each event. \cr
   \code{DT_ENTRY} \tab Start Date of event \cr
   \code{END_DT_ENTRY} \tab Optional end of period to define regimes or ranges of events. \cr
}
}
\examples{
\dontrun{
tail(fg_get_dates_of_interest("fedmoves"),3)
# To add (for example) a new FOMC cut of 50bps on 6/16/2026:
newdoi <-data.table(category="fedmoves",eventid="F:-50",
            DT_ENTRY=as.Date("6/16/2026",format="\%m/\%d/\%Y"))
fg_update_dates_of_interest(newdoi)
tail(fg_get_dates_of_interest("fedmoves"),3)

fg_get_colors("lines")

# To switch to graduated scales, with darkest being first
require(RColorBrewer)
fg_get_colors("lines",n_max=3) -> oldcolors
fg_update_colors( oldcolors[,let(color=  rev(RColorBrewer::brewer.pal(8,"GnBu"))[1:5])][] )
fg_get_colors("lines",n_max=3)

fg_reset_to_default_state("color")
}

}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}
}
\keyword{datasets}
