% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_ggplot.R
\name{fg_eventStudy}
\alias{fg_eventStudy}
\title{Event Studies graphs}
\usage{
fg_eventStudy(indata,dtset,output="path",changeas="diff",
             nbd_back=10,nbd_fwd=20,n_color_switch=5,
             title="Events",maxdelta=+Inf,meltvar="variable",verbose=FALSE)
}
\arguments{
\item{indata}{A data.frame with at least one date column and multiple numeric columns.  If melted,
must also contain the character column specified in parameter \code{meltvar}#'}

\item{dtset}{A list of dates or a \code{data.table} with a column of (event) dates and a character column with unique names for each date.}

\item{output}{(Default \code{path}) : String with type of output desired.  Choices are shown below by category:
\itemize{
\item \code{data},\code{summary},\code{stats} : \code{data.table} with eVent moves by asset and business day relative event, a summary of events by asset and eventid,
or statistics relative to a crossing of events and assets.
\item \code{path}, \code{pathbyvar}, \code{pathbyevent} Show paths of time series moves, by both events and time series, just by time series, or just by event.
\item \code{lmbyvar}, \code{lmbyevent} Show paths of time series moves by time series, or by event, but include linear regression of move ~ time.
\item \code{loessbyvar}, \code{loessbyevent} Show paths of time series moves by time series, or by event, but smoothed loess curves of move ~ time.
\item \code{medbyvar},\code{medbyevent} Median moves by time series or by event.
\item \code{box}, \code{boxbyvar},\code{boxbyevent} Box plots of moves by both events and time series, just by time series, or just by event.
\item \code{scatter} Scatter plot of cumulative move from \code{event-nbd_back} vs \code{event+nbd_fwd}, with medians and regions of movement.
}}

\item{changeas}{(Default \code{diff}) Character string in \code{c("diff","return","returnbps")} describing how changes are displayed.  Log returns are used.}

\item{nbd_back}{(Default 10) Positive integer for number of days prior to event are considered.}

\item{nbd_fwd}{(Default 20) Positive integer for number of days after event are considered.}

\item{n_color_switch}{(Default 5) A positive integer after which colors are displayed as gradients instead of separate colors. See Examples.}

\item{title}{Character string for title of graph}

\item{maxdelta}{(Default \code{+Inf}) Integer to cut off the number of days forward shown, useful if you want to calculate full period statistics.}

\item{meltvar}{(Default \code{variable})  Name of column describing distinct time series if \code{indata} is in long (melted) format,}

\item{verbose}{(Default \code{FALSE})  Print Progress of calculations.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot()}} object with the  events analysis requested by \code{output} parameter, or a \code{data.frame} with statistics if \code{output} in \verb{c("data"","summary","stats")}
}
\description{
Summarizes and plots moves in data from a given set of event dates. Plots are designed to maintain reasonable aesthetics with as
either time series or event dates increase.
}
\details{
Event Studies
}
\examples{
dtset <- fg_get_dates_of_interest("fedmoves",startdt="2024-01-01")[,.(DT_ENTRY,text=eventid2)]
fg_eventStudy(yc_CMSUST,dtset,title="Fed Cuts",output="stats")
fg_eventStudy(yc_CMSUST,dtset,title="Fed Cuts",output="pathbyevent")
fg_eventStudy(yc_CMSUST,dtset,title="Fed Cuts",output="medbyvar")
fg_eventStudy(yc_CMSUST,dtset,title="Fed Cuts",output="lmbyvar",n_color_switch=0)
fg_eventStudy(yc_CMSUST,dtset,nbd_back=3,nbd_fwd=10,title="Fed Cuts",output="boxbyvar")
fg_eventStudy(yc_CMSUST,dtset,title="Fed Cuts",output="scatter")

}
