% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_findTurningPoints}
\alias{fg_findTurningPoints}
\title{Add turning points to time series graphs}
\usage{
fg_findTurningPoints(
  indta,
  rtn = "dates",
  method = "pctchg",
  npts = 10,
  pts_of_interest = "change",
  pctabovemin = 0.05,
  maxwindow = -1,
  addlast = FALSE,
  cpmmethod = "GLR",
  ...
)
}
\arguments{
\item{indta}{Time series \code{data.table} with a date as the first column and a value series as the second column,
or a \code{\link[prophet:prophet]{prophet::prophet()}} object}

\item{rtn}{string with what to return ('dates','data','all')}

\item{method}{string describing method of finding Turning Points
\itemize{
\item \code{"pctchg"} : (Default) Find \code{npts} largest percentage changes with a miniumum window between them
\item \code{"cpm"} : Uaw \link[cpm:cpm-package]{cpm-package} to find change points
}}

\item{npts}{Number of change points to find}

\item{pts_of_interest}{string in 'change' (default) or 'value'}

\item{pctabovemin}{Minimum percentage change to look for.}

\item{maxwindow}{Integer (default -1) which limits (if positive) the minimum number of observations between change points.}

\item{addlast}{Logical (default: FALSE) to add an event with final observation.}

\item{cpmmethod}{String (default: "GLM") passed to \code{\link[cpm:processStream]{cpm::processStream()}}}

\item{...}{Additional parameters passed to \link[cpm:cpm-package]{cpm-package}}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Event Helpers : fg_findTurningPoints
}
\examples{
dta <-eqtypx[,.(date,QQQ,TLT)]
fgts_dygraph(dta,event_ds=fg_findTurningPoints(dta),title="With turningPoints")
}
