% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_helpers.R
\name{fg_predict}
\alias{fg_predict}
\title{Forecast_Helpers}
\usage{
fg_predict(p, confidence = 80, seriesnm = "")
}
\arguments{
\item{p}{Output from \code{\link[forecast:reexports]{forecast::forecast()}}}

\item{confidence}{(Default: 80) Confidence interval (in percent) to display}

\item{seriesnm}{Series name which has been forecast.  Note that \code{\link[forecast:reexports]{forecast::forecast()}} loses the name
of what's been forecast, necessitating this.}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{forecastdataset} parameter
}
\description{
\code{fg_predict} Converts a \code{\link[forecast:reexports]{forecast::forecast()}} output into \code{\link[=fgts_dygraph]{fgts_dygraph()}} \code{forecastdataset} format.
}
\examples{
require(forecast)
require(zoo)
px_small <- narrowbydtstr(eqtypx,"-2y::")
fcst_one <- function(ticker) {
  t1_ts <- zoo::zoo(px_small[[ticker]],px_small[["date"]])
  forecast::ets(t1_ts) |> forecast::forecast(h=36) |>  fg_predict(seriesnm=ticker)
  }
fpred <- merge(fcst_one("QQQ"),fcst_one("IBM"),by="date")
fgts_dygraph(px_small,title="With Forecasts", roller=1,dtstartfrac=0.7,
              forecast_ds=fpred,verbose=TRUE)

}
