% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_constants.R
\name{fg_replace_theme}
\alias{fg_replace_theme}
\title{Theme Management}
\usage{
fg_replace_theme(newTheme, persist = TRUE)
}
\arguments{
\item{newTheme}{A new ggplot2 theme}

\item{persist}{(Default: TRUE) Keep changes across invocations of the package.}
}
\value{
Acknowledgement message if saved
}
\description{
Sets, gets, or resets default ggplot themes to be used.
}
\examples{
require(ggplot2)
fg_replace_theme(ggplot2::theme_dark(),persist=FALSE)
}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}, \code{\link[=fg_scatplot]{fg_scatplot()}}
}
