% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat_ggplot.R
\name{fg_scatplot}
\alias{fg_scatplot}
\title{Flexible and easy  scatterplot generator, with time specific enhancements}
\usage{
fg_scatplot(indata,plotform,type="scatter",datecuts=c(7,66),
               noscales="",xdecoration="",ydecoration="",annotatecorners="",
               tsize=3,psize=1,n_color_switch=7,n_hex_switch=400,repel=TRUE,jitter=c(0,0),
               title="",subtitle="",caption="",axislabels="",
               boundbox=c(),boundboxtype="",gridstyle=NA_character_,legendinside=FALSE,
               tformula=formula("y~x"),returnregresults=FALSE,
               keepcols="", meltvar="variable",melted=NULL)
}
\arguments{
\item{indata}{\code{data.frame} with columns for (x,y) coordinates and possibly other categorical data or a date column. Alternatively,
\code{indata} can be in long format with \code{meltvar} present.  Note that aesthetic characteristics (if used) must be present for both
long and wide input formats.}

\item{plotform}{A text formula describing how to set up the graph.  The formula is of the
general form

\verb{y ~ x + option:<column_name>,<aesthetic category> + ...}

where \code{y} is plotted against \code{x} and aesthetics for each point are controlled by one or more \code{option} clauses each followed
by zero or more optional parameters.   If the option applies to all points then the first parameter \code{column_name} must be in \code{indata}.
By default, points or symbols are plotted.  By general category, the options are

Aesthetic options:
\itemize{
\item \verb{color:column<,aes_set>} sets the color of each point or label by data in \code{column}
\item \verb{symbol:column<,aes_set>} sets the symbol or shape cof each point or label by data in \code{column}
\item \verb{size:column<,aes_set>} sets the size of each point by the data in \code{column}
}

Date specific options:
\itemize{
\item \code{doi:recent} partitions data by number of days in \code{datecuts}prior to the last day in \code{indata}
\item \verb{doi:<doiset>} partitions data by date ranges obtained from dates of interest set \verb{<doiset>}.  See \code{\link[=fg_get_dates_of_interest]{fg_get_dates_of_interest()}}
\item \verb{point:<value|label|anno><all>} adds highlights for either the last date in \code{indata} or the last
date for each group (if \code{all}). \code{value} gives coordinates, \code{label} the label in the \code{color} column, while \code{anno}
adds lines to each axis.
}

Text options:
\itemize{
\item \verb{[text|label|labelhilight|tooltips]:column<,aes_set>} : Plots the text in character \code{column} as text (without border),
label (with border), filled in label, or mouse-over tooltip. (See details)
}

Other annotations:
\itemize{
\item \code{ellipse} adds an ellipse around the points using \code{\link[ggplot2:stat_ellipse]{ggplot2::stat_ellipse()}}
\item \verb{hull<:quantile>} draws a convext hull around points with \verb{<quantile>} (default 0) points removed. (See details)
\item \verb{xline<:level=0>} draws a vertical line  at \code{level}
\item \verb{yline<:level=0>} draws a horizontal line  at \code{level}
\item \verb{grid:<dotted|dotted_x|dotted_y|none>} formats background grids
}}

\item{type}{character string for the type of graph to plot:
\itemize{
\item \code{scat} plots points, text or labels
\item \verb{lm<one><noeqn><nofill>} adds linear regression lines per \code{color} category or across all points (\code{lmone}).
\item \verb{loess<one><noeqn>}adds loess line per per \code{color} category or across all points (\code{loessone})
\item \code{density} Creates a density plot
}

if \code{noeqn} is part of the string, equations are suppressed.  If \code{one} is part of the string, no subcategories are used. `nofill``
removes confidence bands.}

\item{datecuts}{list of integers (Default \code{c(7,66)} for days prior to last date to make date classes. (See examples and \code{doi:recent} as above.)}

\item{noscales}{String to suppress guides with any of \verb{<color|size|symbol>}}

\item{xdecoration, ydecoration}{2 element string list to add to either side of an axis label.}

\item{annotatecorners}{4 element string list to add notes to each of 4 quadrants of the graph.  See examples.}

\item{tsize}{default text size (with some scaled variations for graph parts such as titles)}

\item{psize}{default point size.}

\item{n_color_switch}{(Default 7): Number of distint color categories beyond which colors are taken from gradient scales,
unless a color set is specified in a \code{color} part of \code{plotform}}

\item{n_hex_switch}{(Default 400): Number of data points beyond which points are replaced with binned hexagons (see \code{\link[ggplot2:geom_hex]{geom_hex()}})}

\item{repel}{(Default TRUE) Text and labels are plotted using \code{\link[ggrepel:ggrepel]{ggrepel()}}}

\item{jitter}{Jitter parameters used by \code{\link[ggplot2:geom_text]{geom_text()}} or \code{\link[ggplot2:geom_label]{geom_label()}} if \code{repel=FALSE}. Default is no jitter.}

\item{title}{Title to add to graph}

\item{subtitle}{Subtitle to add to graph}

\item{caption}{Caption to add to graph}

\item{axislabels}{Semicolon separated string with x and y labels, e.g \verb{date;OAS}}

\item{boundboxtype, boundbox}{string describing how to use bounding boxes.  If \code{"identify"} is part of the string, then data
is truncated to the calculated bounding box and notations to that effect are added to the graph.  If not, then points
outside the box are dropped.
\itemize{
\item \code{prob|probidentify} calculates bounding boxes from quantilies of \code{x} and \code{y} data.  See vignette for details.
\item \code{value|valueidentify} are minimum and maximum values of \code{x} and \code{y} axes to show.
If \code{boundbox} is a list of two numbers, the y axis is truncated to those values,
If \code{boundbox} is a list of 4 numbers \code{c(xmin,xmax,ymin,ymax)} data is truncated to that box.
}}

\item{gridstyle}{String in \verb{<dotted|dotted_x|dotted_y|none>} to contol grids as in \code{grid} option above.}

\item{legendinside}{(Default: TRUE) Put all guides inside the graph.}

\item{tformula}{(Default \code{y~x}) Formula used within \code{lm} or \code{loess} stats .}

\item{returnregresults}{Return a two element list c(plot,regression \code{data.frame}). Only available for linear models, and uses the first amoung
options \code{c("color","symbol","size","alpha")} as grouping variables}

\item{keepcols}{list of \code{indata} columns to be kept with the graph data, useful for further faceting using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{meltvar}{(Default \code{"variable"}) If \code{indata} is melted, then this is used to create \code{x} and \code{y} categories.}

\item{melted}{(Default:NULL) If \code{FALSE} forces data not to be melted if \code{meltvar} in \code{indata}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object with desired graph, or a \code{\link[ggiraph:girafe]{ggiraph::girafe()}} object if \code{tooltips} is in the \code{plotform} string.
}
\description{
Plots bivariate plots with some time-series specific enhancements.  Rather than programmatically describing
graph aesthetics, a simple formula-based approach is used.  This approach allows quick specification of
many customizaton options.
}
\details{
fg_scatplot:  Easy scatterplot generator, with time specific enhancements

\code{indata} can either be in wide ('date' ,'series1',...) format or normalized (long) format
('date','variable','value',...) format.  This package infers date columns names from column types and casts or
pivot_wider to get \code{x} and \code{y} columns.  Note that aesthetic characteristics (if used) must be present for both
long and wide input formats.

Default aesehetic sets used for portions of the graph each have their own names, which can be seeing by running
\code{\link[=fg_print_aes_list]{fg_print_aes_list()}} and modified or added to using \code{\link[=fg_update_aes]{fg_update_aes()}}.  The default theme can be modified using
\code{\link[=fg_replace_theme]{fg_replace_theme()}}.  Both aesthetic changes and theme changes are persistent across R sessions.

Use of \verb{doi:} in \code{plotform} string supercedes color aesthetics otherwise specified.  (Is this true?)

If \code{tooltips} are used, the result of \code{fg_scatplot} must be viewed using \code{print(girafe(ggobj=fg_scatplot(...)))}
See \code{\link[ggiraph:geom_point_interactive]{ggiraph::geom_point_interactive()}}

Winsorized hulls with quantile cutoff \code{q} are formed using the closest (by euclidean distance) \code{1-q} points to the
geograpic center of the entire set.

Captions are added if data is truncated or omitted by the bounding box procedure.
}
\examples{
# Simple text examples
require(data.table)
dt_mtcars=data.table(datasets::mtcars)
dt_mtcars$id=lapply(rownames(datasets::mtcars),\(x) last(strsplit(x," ")[[1]]))
fg_scatplot(dt_mtcars,"disp ~ hp + color:am + text:id","scatter",title="text basic")
fg_scatplot(dt_mtcars,"disp ~ hp + color:carb + label:id","scatter",
                n_color_switch=0,title="scat color switch")
# Plotting data with dates:
set.seed(1); ndates <- 400;  dlyvol <- 0.2/sqrt(260)
rtns <- cbind(cumsum(rnorm(ndates,sd=dlyvol)),cumsum(rnorm(ndates,sd=dlyvol)))
dttest <- data.table(date=seq(as.Date("2021-01-01"),as.Date("2021-01-01")+ndates-1),
            xtest=100*(1+rtns[,1]),ytest=100*(1+rtns[,2]),
            ccat=fifelse(runif(ndates)<=0.2,"Rare","mkt"))
# Making categories out of recent data
fg_scatplot(dttest,"ytest ~ xtest + doi:recent","scatter",datecuts=c(66,122),title="from recent")
fg_scatplot(dttest ,"ytest ~ xtest + color:ccat + doi:recent + point:label","scat",
              datecuts=c(7,66),title="recent w label")
# Makes categories out of event sets from [fg_get_dates_of_interest()]
fg_scatplot(dttest,"ytest ~ xtest + doi:regm","scatter",title="from a regime")
# Point graphing switches.
fg_scatplot(dttest,"ytest ~ xtest + color:ccat","lm",n_hex_switch=100,title="Hex Switch")
# Quick changes to aesthetic sets
fg_scatplot(dttest,"ytest ~ xtest + color:ccat,altlines_6","loess",title="Alternate colors")
# Extra summarizatons
fg_scatplot(dttest,"ytest ~ xtest + color:ccat + hull:0.1 + ellipse","lm",title="Curves")
# Annotations
fg_scatplot(dttest,"ytest ~ xtest + color:ccat + point:labelall","scat",title="Last Values")
fg_scatplot(dttest,"ytest ~ xtest + color:ccat + point:anno","scat",annotatecorners="NW;NE;SE;SW",
                legendinside = FALSE)
}
