% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_ggplot.R
\name{fg_seasonalstudy}
\alias{fg_seasonalstudy}
\title{Flexible Seasonality Graphs}
\usage{
fg_seasonalstudy(
  indta,
  seasonaltype = "yr",
  seasonaldateset = NULL,
  day_offset = 0,
  bdaysonly = TRUE,
  graphtype = "flex",
  normalize = "",
  projectfwd = "",
  projfwd_wt = 0.9,
  yvar = "value",
  title = NULL,
  yrange = NULL,
  n_color_switch = 7,
  n_hex_switch = 20,
  line_aes_set = "lines",
  line_on_lastdate = TRUE,
  killbad_eop = FALSE,
  return_dates = FALSE
)
}
\arguments{
\item{indta}{A data.frame with at least one date column and a numeric column with the name given by \code{yvar}}

\item{seasonaltype}{(default: \code{"yr"}) A string denoting the periodicity of the analysis, must be in one of
\verb{c("yr","qtr","mo","wk",IMMroll","optmo","optqtr")}.  Not used if \code{seasonaldataset} is specified instead}

\item{seasonaldateset}{(default:NULL) An optional dataset with two columns: A Date column defining each period, and possibly a character
variable with a period identifier.}

\item{day_offset}{(default: 0) Number of days (determined by bdaysonly) to offset each period.  Applies only to \code{c("yr","qtr","mo","wk")}}

\item{bdaysonly}{(default:TRUE)  Only consider NYSE business days.}

\item{graphtype}{(default: "flex")  One of the following\tabular{cl}{
   \code{graphtype} \tab Description \cr
   \code{line} \tab A line for every period, with identifiers placed near the end \cr
   \code{hex} \tab A density plot for each day of the seasonal period, with the last seasonal  period  kept as a line \cr
   \code{flex} \tab A line graph if there are less than \code{n_hex_switch} seasonal periodss \cr
   \code{stat} \tab A line graph showing 10tyh, 50th, and 90th percentiles of values for each day in a seasonal period \cr
}}

\item{normalize}{(default "").  How to normalize each period.  Default is no normalization.  Other options are\tabular{cl}{
   \code{normalize} \tab Description \cr
   \code{relative} \tab Older seasonal periods adjusted to match beginning of latest  period \cr
   \code{index} \tab Each seasonal period is expressed as index from beginning of each period \cr
}}

\item{projectfwd}{(default "")  Project forward based on same number of days in seaonal period.  Default is no projection.\tabular{cl}{
   \code{normalize} \tab Description \cr
   \code{mean} \tab Use mean (by days in seasonal period) of each cumulative percentage change from the start of the period \cr
   \code{weighted} \tab Weight previous period cumulative percantge change with an expoential decay using \code{projfwd_wt} ** (Periods back) \cr
}}

\item{projfwd_wt}{See above}

\item{yvar}{(default:\code{"value"}) Series to use in \code{indta}}

\item{title}{(default: NULL) Title for graph.}

\item{yrange}{(default: NULL) y axis Range for which to focus data.}

\item{n_color_switch}{(default 7) Number of periods past which lines will be colored by a descending scale.}

\item{n_hex_switch}{(default 20) Number of periods past which a \code{hex} graph will be used, if \code{flex} is chosen above.}

\item{line_aes_set}{(default \code{"lines"}) Aes color set for discrete lines.}

\item{line_on_lastdate}{(default TRUE) Add a vertical line at the last observations day in period.}

\item{killbad_eop}{(default FALSE):  DO not show periods for which there are at least 60\% of the mean number of observations
per day of the seasonal period.  If used, this helps to curtail  extreme moves at the end of a period.  For example,
this would redact the 366th day of the year.}

\item{return_dates}{(default: FALSE)  Return \code{list(graph,dates)} instead of just the graph.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot()}} object displaying seasonality
}
\description{
Flexible and general seasonality graphs.
}
\details{
Seasonality Graphs
}
\examples{
require(data.table)
assetcols <- c("EEM","IBM","QQQ","TLT")
eqtyidx<-eqtyrtn[,(assetcols):=lapply(.SD,\(x) 100*(exp(cumsum(fcoalesce(x,0))))),
            .SDcols=assetcols]
fg_seasonalstudy(eqtyidx,yvar="TLT",seasonaltype="yr",normalize="index",projectfwd="mean")
# Earnings seasons
earnings_dates <- earnings_ibm[,.(reportedDate,divdt=format(reportedDate,"\%Y\%m\%d"))]
fg_seasonalstudy(eqtyidx,yvar="IBM",seasonaldateset = earnings_dates,graphtype="stat",
   normalize="index",projectfwd="mean")
}
