% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_signal_to_events}
\alias{fg_signal_to_events}
\title{Add single date signals to time series graphs.}
\usage{
fg_signal_to_events(signal_df, colormap)
}
\arguments{
\item{signal_df}{A two-column \code{data.frame} with first being a date and second being any (factor-like) signal parameter.}

\item{colormap}{A two column \code{data.frame} with the first being the possible signal (see Example) and the second a color.
description}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Event Helpers: fg_signal_to_events
}
\details{
This helper applies run-length encoding to match the signal in \code{signal_df} to the color in \code{colormap}
}
\examples{
# A simple moving average strategy with threshold
require(data.table)
ma_signal<-eqtypx[,.(date,sig=cut(frollmean(EEM,5)-frollmean(EEM,20),
                    c(-10,-0.5,0.5,10),labels=c("long","flat","short")),EEM)]
colormap<-data.frame(sig=c("long","flat","short"),color=c("#f56462","white","#6161ff"))
fgts_dygraph(eqtypx[,.(date,EEM)],event_ds=fg_signal_to_events(ma_signal,colormap),
    dtstartfrac=0.8,roller=1,title="5/20 MA positions")
}
