% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_helpers.R
\name{fg_sweep}
\alias{fg_sweep}
\title{Add forecasts from the sweep package}
\usage{
fg_sweep(swept_data, confidence = 80)
}
\arguments{
\item{swept_data}{Data resulting from a \code{\link[sweep:sw_sweep]{sweep::sw_sweep()}} call}

\item{confidence}{(Default: 80) Confidence interval (in percent) to display}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{forecastdataset} parameter
}
\description{
\code{fg_sweep} Augments a \code{\link[sweep:sw_sweep]{sweep::sw_sweep()}} output into \code{\link[=fgts_dygraph]{fgts_dygraph()}} \code{forecastdataset} format.  See
\href{https://business-science.github.io/sweep/articles/SW00_Introduction_to_sweep.html}{Introduction to Sweep}
}
\details{
Forecast Helpers
}
\examples{
if (
    requireNamespace("forecast", quietly = TRUE) &
    requireNamespace("timetk", quietly = TRUE) &
    requireNamespace("sweep", quietly = TRUE)
) {
fcst_eqtypx <- timetk::tk_ts(eqtypx[,.(date,QQQ)]) |> forecast::ets() |>
      forecast::forecast(h=30) |> sweep::sw_sweep(timetk_idx=TRUE)
fcst_in <- fg_sweep(fcst_eqtypx)
toplot <- eqtypx[,.(date,IBM,QQQ)]
fgts_dygraph(toplot,title="With Forecasts", roller=1,dtstartfrac=0.7,forecast_ds=fcst_in)
}
}
