% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_tq_divs}
\alias{fg_tq_divs}
\title{Event Helpers: fg_tq_divs}
\usage{
fg_tq_divs(tickers, divs_ds = NULL, ticker_in_label = TRUE)
}
\arguments{
\item{tickers}{List of tickers to get dividends for.}

\item{divs_ds}{Alternatively a \code{data.frame} previously obtained using \code{\link[tidyquant:tq_get]{tidyquant::tq_get()}} with columns (\code{symbol},\code{date},\code{value})}

\item{ticker_in_label}{(Default: TRUE) Make label ticker and the dividend.}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Calls \code{\link[tidyquant:tq_get]{tidyquant::tq_get()}} to get dividends for a given set of tickers.  A previously created \code{data.frame} can also be input.
}
\examples{
require(tidyquant)
fgts_dygraph(eqtypx,title="With divs",dtstartfrac=0.8,event_ds=fg_tq_divs(c("IBM","QQQ")))

}
