% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_constants.R
\name{fg_update_colors}
\alias{fg_update_colors}
\alias{fg_get_colors}
\alias{fg_update_line_colors}
\alias{fg_reset_to_default_state}
\title{Maintain Colors}
\usage{
fg_get_colors(item = "", n_max = NA_integer_)

fg_update_colors(indta, replace = FALSE, persist = TRUE)

fg_update_line_colors(colorlist, replace = FALSE, persist = TRUE)

fg_reset_to_default_state(reset = "all")
}
\arguments{
\item{item}{(Default: "") A grep string for categories desired.}

\item{n_max}{Maximum number of rows to return.}

\item{indta}{\code{data.table} with columns as shown in details.}

\item{replace}{(Default: FALSE) Replaces existing dates of interest with new set provided, otherwise replaces/inserts new rows only.}

\item{persist}{(Default: TRUE) Keep changes across invocations of the package.}

\item{colorlist}{List with up to 14 new colors just for line (series) coloring}

\item{reset}{(Default: "all"), options in ("all","colors","doi") to reset to defaults with the package.}
}
\value{
\code{data.frame} or message
}
\description{
\code{fg_get_colors()} gets default color sets for graphs
\code{fg_update_colors()} updates or replaces default colors
\code{fg_update_line_colors()} replaces line colors only
\code{fg_reset_to_default_state()} resets colors and/or dates of interest
}
\details{
New data passed into \code{\link[=fg_update_colors]{fg_update_colors()}} persists across future loads of the package unless \code{persist=FALSE}.
New color datasets must have at least three columns:\tabular{ll}{
   Column \tab Meaning \cr
   \code{category} \tab Coloring category, typically \code{"lines"} for line colors. \cr
   \code{variable} \tab Any string that can be sorted or grepped to map to data. \cr
   \code{color} \tab Color desired \cr
}
}
\examples{
\dontrun{
fg_get_colors("lines")
# To switch to graduated scales, with darkest being first
require(RColorBrewer)
fg_get_colors("lines",n_max=3) -> oldcolors
fg_update_colors( oldcolors[,let(color=  rev(RColorBrewer::brewer.pal(8,"GnBu"))[1:3])][] )
fg_get_colors("lines",n_max=3)
fg_update_line_colors(c("black","green","red","blue"),persist=FALSE)
fg_get_colors("lines",n_max=4)
fg_reset_to_default_state("color")
}

}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}
}
