% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinanceGraphs-package.R
\docType{package}
\name{FinanceGraphs-package}
\alias{FinanceGraphs}
\alias{FinanceGraphs-package}
\title{FinanceGraphs: Flexible Graphs for Analysis of Financial Data and Time Series}
\description{
Flexible wrappers around R graphics modules 'dygraphs' \url{https://dygraphs.com/} and 'ggplot2' \url{https://ggplot2.tidyverse.org/} to visualize data commonly found in Financial Studies, with an emphasis on time series. Interactive time series plots include multiple options for incorporating external data such as forecasts and events. Other static plots useful for time series data include an intuitive and generic scatter plotter, a boxplot generator suitable for multiple time series, and event study plotters for time series analysis around sets of dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/derekholmes0/FinanceGraphs}
  \item Report bugs at \url{https://github.com/derekholmes0/FinanceGraphs/issues}
}

}
\author{
\strong{Maintainer}: Derek Holmes \email{derek@derekholmes.com} [copyright holder]

}
\keyword{internal}
