% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gendtstr}
\alias{gendtstr}
\title{Form exact datestring from relative datestring}
\usage{
gendtstr(x, today = Sys.Date(), rtn = "dtstr")
}
\arguments{
\item{x}{String describing generalized date as of today}

\item{today}{Default \code{Sys.Date()}}

\item{rtn}{What to return, Default is string of form \code{yyyy-mm-dd::yyyy-mm-dd}.  Can also be list\verb{, }first\verb{, }days`}
}
\value{
depends on value of \code{rtn}
}
\description{
Form exact datestring from relative datestring
}
\examples{
gendtstr("-3m::")

}
