% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{imfdta}
\alias{imfdta}
\title{IMF Economic FOrecasts}
\format{
\subsection{\code{imfdta}}{

A long format \code{data.table} with both historical economic data and projections.
\describe{
\item{CC}{ISO Country code}
\item{SUBJ}{IMF subject classification}
\item{value}{Value of historical data or forecast}
\item{variable}{Short abbreviation for SUBJ}
\item{ctryname}{Full country name correponding to CC}
\item{region}{Investment Region each country is in}
\item{usedccy}{Dominant currency used in Country}
\item{date}{Date of historical data or of forecast}
}
}
}
\source{
\url{https://www.imf.org/en/publications/weo}
}
\usage{
imfdta
}
\description{
IMF World Economic Outlook June 2025
}
\keyword{datasets}
