% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{narrowbydtstr}
\alias{narrowbydtstr}
\title{Narrow dataset by date string}
\usage{
narrowbydtstr(
  xin,
  dtstr = "",
  includetoday = TRUE,
  windowdays = 0,
  invert = FALSE,
  addindicator = FALSE
)
}
\arguments{
\item{xin}{INput \code{data.frame} or \code{data.table}}

\item{dtstr}{Generalized Date string of the form \verb{<yyyy-mm-dd>::<yyyy-mm-dd>} or e.g. \verb{-3m::}}

\item{includetoday}{(Default: TRUE) pass either today \code{Sys.Date()} or \code{Sys.Date()-1} to \code{gendtstr}}

\item{windowdays}{(Default: 0)Number of additional days to add at beginning of series}

\item{invert}{(Default: FALSE) Return dates not in \code{dtstr}}

\item{addindicator}{(Default: FALSE) Returns original dataset with logical variable \code{inrange} if date is in desired range.}
}
\description{
Narrow dataset by date string
}
\examples{
narrowbydtstr(eqtypx,"-2m::-1m")

}
