% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{nomfxdta}
\alias{nomfxdta}
\title{Nominal FX levels}
\format{
\subsection{\code{nomfxdta}}{

A tibble with 13134 observations and 3 columns
\describe{
\item{date}{Date Of Observation}
\item{variable}{CUrrency}
\item{value}{Currency/USD}
}
}
}
\source{
\url{https://business-science.github.io/tidyquant/}
}
\usage{
nomfxdta
}
\description{
Closing Nominal currency levels (local currency per dollar)
}
\keyword{datasets}
