% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{overlay_eventset}
\alias{overlay_eventset}
\title{Event Helpers: overlay_eventset}
\usage{
overlay_eventset(indta, ncutsperside = 4, center = 0, invert = FALSE)
}
\arguments{
\item{indta}{Time series \code{data.table} with a date as the first column and a value series as the second column.}

\item{ncutsperside}{: Integer with number of colors to use on each side of 'center'}

\item{center}{: String or Double as follows:
\itemize{
\item Double (default 0) Normalize data by subtracting \code{center}
\item "median" Normalize data by subtracting median of all observations.
\item "zscore" Normalize data by using standard \code{\link[=scale]{scale()}} function
}}

\item{invert}{Use opposite color schemes for data, i.e. "red" for good outcomes}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{eventdataset} parameter
}
\description{
Event Helpers: overlay_eventset
}
\details{
Always uses first date column and first numeric columns in data.  If \code{indta} has multiple series, filter them before calling the function.
}
\examples{
overlay_eventset(consumer_sent,center="zscore")

}
