% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ratings_db}
\alias{ratings_db}
\title{Ratings Database}
\format{
\subsection{\code{ratings_db}}{

A tibble with 110 observations and 5 columns
\describe{
\item{CREDIT}{Short Credit name for a few countries}
\item{AGENCY}{Ratings Agency}
\item{RATING}{Rating code for each country specific to each agency}
\item{WATCH}{Character indicator if ratings is watch positive or negative}
\item{DT_ENTRY}{Date of ratings annoucement}
}
}
}
\source{
\url{https://ratingshistory.info/}
}
\usage{
ratings_db
}
\description{
Ratings changes for a few select sovereigns.
}
\keyword{datasets}
