% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{recession_indic}
\alias{recession_indic}
\title{Monthly recession indicator, FRED code RECPROUSM156N}
\format{
\subsection{\code{recession_indic }}{

A data frame with 120 rows and 3 columns
\describe{
\item{symbol}{FRED identifier}
\item{date}{Date of report}
\item{pricce}{Observation}
}
}
}
\source{
\url{https://fred.stlouisfed.org/}
}
\usage{
recession_indic
}
\description{
Monthly recession indicator, FRED code RECPROUSM156N
}
\keyword{datasets}
