% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{earnings_ibm}
\alias{earnings_ibm}
\title{IBM Earnings}
\format{
\subsection{\code{earnings_ibm}}{

A \code{data.table} with 120 observations and e columns
\describe{
\item{reportedDate}{Earnings announcement date}
\item{reportedEPS}{Reported Earnings per Share}
\item{symbol}{Company}
}
}
}
\source{
\url{https://alphavantage.co}
}
\usage{
earnings_ibm
}
\description{
IBM Earnings download
}
\keyword{datasets}
