% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{eqtypx}
\alias{eqtypx}
\title{Equity Prices}
\format{
\subsection{\code{eqtypx}}{

A data table with 2529 obersavatin and 5 variables
\describe{
\item{date}{Date Of Observation}
\item{EEM}{EEM Closing Price}
\item{IBM }{IBM  Closing Price}
\item{QQQ}{QQQ Closing Price}
\item{TLT}{TLT Closing Price}
}
}
}
\source{
\url{https://finance.yahoo.com}
}
\usage{
eqtypx
}
\description{
Closing Equity Prices
}
\keyword{datasets}
