% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{eqtyrtn}
\alias{eqtyrtn}
\title{Equity returns}
\format{
\subsection{\code{eqtyrtn}}{

A tibble with 2529 observations and 6 columns
\describe{
\item{date}{Date Of Observation}
\item{EEM}{EEM Log daily return}
\item{IBM }{IBM Log daily return}
\item{QQQ}{QQQ Log daily return}
\item{TLT}{TLT Log daily return}
\item{p_TLT_QQQ}{ROlling 66 business day regression of TLT on QQQ p.value}
\item{ibmdiv}{Dividends reported for IBM}
}
}
}
\source{
\url{https://business-science.github.io/tidyquant/}
}
\usage{
eqtyrtn
}
\description{
Closing Equity returns and a rolling regression
}
\keyword{datasets}
