% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_ratingsEvents}
\alias{fg_ratingsEvents}
\title{Event Helpers : fg_ratingsEvents}
\usage{
fg_ratingsEvents(credit, ratings_db, agency = "S.P")
}
\arguments{
\item{credit}{String with name of credit to look up in 'ratings_db'}

\item{ratings_db}{A 'data.table' or 'data.frame' with the all of the following columns:
| column | type | description |
|: --- |: --- :|: --- |
| CREDIT | character | Name of credit  |
| AGENCY | character | Name of ratings agency  |
| RATING   | character | Rating assigned  |
| WATCH   | character | Watch denoted by anything with "+" or "-" in the string |
| DT_ENTRY  | Date | Date which ratings or ratings change was issued |}

\item{agency}{String (default 'S.P') with 'AGENCY to look up in 'ratings_db'}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Event Helpers : fg_ratingsEvents
}
\details{
Investment grade ratings are shaded in blue, High Yield are in red. Darker areas are closest to the cutoff between the two.
}
\examples{
copdta <- nomfxdta |> dplyr::filter(variable=="COP")
fgts_dygraph(copdta,title="COP with Ratings",dtstartfrac=0.3,
        event_ds=fg_ratingsEvents("COLOM",ratings_db,agency="S.P"))

}
