% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_fcst_set}
\alias{example_fcst_set}
\title{Example forecasts}
\format{
\subsection{\code{example_fcst_set}}{

A data table with forecasts for two equities
\describe{
\item{date}{Date}
\item{QQQ.f}{Forecast price for QQQ}
\item{QQQ.flo}{80th percentile in Forecast CI for QQQ}
\item{QQQ.fhi}{20th percentile in Forecast CI for QQQ}
\item{IBM.f}{Forecast price for IBM}
\item{IBM.flo}{80th percentile in Forecast CI for IBM}
\item{IBM.fhi}{20th percentile in Forecast CI for IBM}
}
}
}
\source{
\url{https://finance.yahoo.com}
}
\usage{
example_fcst_set
}
\description{
IBM Stock price forecasts from ets baseline ets model
}
\keyword{datasets}
