% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_fcst_set}
\alias{example_fcst_set}
\title{Example forecasts}
\format{
\subsection{\code{example_fcst_set}}{

A named list of two outputs from forecasts on equity stock prices.
\describe{
\item{Point Forecast}{Forecast}
\item{Lo 80}{80th percentile lower confidence bound}
\item{Lo 95}{95th percentile lower confidence bound}
\item{Hi 80}{80th percentile upper confidence bound}
\item{Hi 95}{95th percentile upper confidence bound}
}
}
}
\source{
\url{https://finance.yahoo.com}
}
\usage{
example_fcst_set
}
\description{
IBM Stock price forecasts from ets baseline ets model
}
\keyword{datasets}
