% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_addbreakouts}
\alias{fg_addbreakouts}
\title{Event_Helpers}
\usage{
fg_addbreakouts(indta, annotationstyle = "singleasdate")
}
\arguments{
\item{indta}{Time series \code{data.table} with a date as the first column and a value series as the second column.}

\item{annotationstyle}{String in set (\code{singleasdate},\code{singleasvalue},'breakno')}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Event Helpers : fg_addbreakouts
}
\examples{
dta <- eqtypx[,.(date,QQQ,TLT)]
fgts_dygraph(dta,event_ds=fg_addbreakouts(dta),title="With Breakouts")

}
