% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_helpers.R
\name{fg_prophet}
\alias{fg_prophet}
\title{Forecast_Helpers}
\usage{
fg_prophet(prophet_data, seriesname = "y")
}
\arguments{
\item{prophet_data}{Data resulting from a \code{\link[prophet:predict.prophet]{prophet::predict.prophet()}}  call}

\item{seriesname}{(Default: \code{"y"}) Series name to attach forecast to}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{forecastdataset} parameter
}
\description{
\code{fg_prophet} Augments a \code{\link[prophet:predict.prophet]{prophet::predict.prophet()}}  output into \code{\link[=fgts_dygraph]{fgts_dygraph()}} \code{forecastdataset} format.
}
\details{
Note that  \code{\link[prophet:predict.prophet]{prophet::predict.prophet()}} loses the name of the series, the
}
\examples{
require(prophet)
p_model <- eqtypx[,.(ds=date,y=QQQ)] |> prophet::prophet()
p_fcst <- predict(p_model,prophet::make_future_dataframe(p_model,periods=60))
fgts_dygraph(eqtypx[,.(date,QQQ)],title="With Prophet Forecasts", roller=1,dtstartfrac=0.6,
      event_ds = fg_findTurningPoints(p_model),
      forecast_ds=fg_prophet(p_fcst,seriesname="QQQ"))

}
