% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_helpers.R
\name{fg_RegimeChange}
\alias{fg_RegimeChange}
\title{Event_Helpers}
\usage{
fg_RegimeChange(indta, usereturns = TRUE, series = NULL, ...)
}
\arguments{
\item{indta}{Time series \code{data.table} with a date as the first column and a value series as the second column.}

\item{usereturns}{(default TRUE) LOgical to take log returns before changepoint calculations.}

\item{series}{Which series in \code{indta} to apply changepoints to}

\item{...}{Parameters passed to \code{\link[RegimeChange:detect_regimes]{RegimeChange::detect_regimes()}}}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{event_ds} parameter
}
\description{
Wrapper around the function \code{\link[RegimeChange:detect_regimes()]{RegimeChange::detect_regimes()()}} to create events for \code{\link[=fgts_dygraph]{fgts_dygraph()}}
}
\details{
Event Helpers : fg_RegimeChange
https://cran.r-project.org/web/packages/RegimeChange/vignettes/introduction.html
}
\examples{
if (requireNamespace("RegimeChange", quietly = TRUE)) {
dta <- tail(eqtypx[,.(date,QQQ,TLT)],260)
eventdt = fg_RegimeChange(dta)
fgts_dygraph(dta,event_ds=eventdt,title="With Breakouts")
}
}
