% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_helpers.R
\name{fg_sweep}
\alias{fg_sweep}
\title{Forecast_Helpers}
\usage{
fg_sweep(swept_data, confidence = 80)
}
\arguments{
\item{swept_data}{Data resulting from a \code{\link[sweep:sw_sweep]{sweep::sw_sweep()}} call}

\item{confidence}{(Default: 80) Confidence interval (in percent) to display}
}
\value{
\code{data.table} suitable for passing into \code{\link[=fgts_dygraph]{fgts_dygraph()}} via the \code{forecastdataset} parameter
}
\description{
\code{fg_sweep} Augments a \code{\link[sweep:sw_sweep]{sweep::sw_sweep()}} output into \code{\link[=fgts_dygraph]{fgts_dygraph()}} \code{forecastdataset} format.  See
\href{https://business-science.github.io/sweep/articles/SW00_Introduction_to_sweep.html}{Introduction to Sweep}
}
\details{
Forecast Helpers
}
\examples{
require(timetk)
require(forecast)
require(sweep)
dta <- eqtypx[,.(date,QQQ)]
fcst_eqtypx <- tk_ts(dta) |> ets() |> forecast::forecast(h=30) |> sw_sweep(timetk_idx=TRUE)
fcst_in <- fg_sweep(fcst_eqtypx)

}
