% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_constants.R
\name{fg_sync_group}
\alias{fg_sync_group}
\title{Date Synchronization}
\usage{
fg_sync_group(gpname = "")
}
\arguments{
\item{gpname}{A string or NULL
\itemize{
\item \code{gpname=NULL} turns of dygraphs synchronization.
\item \verb{gpname=<string>} set the common group to \verb{<string>}
\item \code{gpname=""} (Default), just returns the current common group name.
}}
}
\value{
current groupname
}
\description{
Sets, gets, or resets a common name to be passed into  \code{\link[=fgts_dygraph]{fgts_dygraph()}} for synchronization.
}
\details{
Use thie to set a common groupname for time scale synchronization (for Markdown or shiny apps),  Only set it in the beginning,
or when needed, and call with NULL to turn synchronization off.
}
\examples{
fg_sync_group()
fg_sync_group("common")
fg_sync_group()
fg_sync_group(NULL)

}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}
}
