% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ggplot.R
\name{fg_tsboxplot}
\alias{fg_tsboxplot}
\title{fg_tsboxplot:  Boxplots of time series}
\usage{
fg_tsboxplot(indt,title="",xlab="",ylab="",
     breaks=c(7,30,90,360), doi="last", normalize="", orderby="",
     boxtype= "",
     dropset="",hilightcats="",
     addline="", #last/mean
     facetform="",
     ycoord=NULL,trimpctile=0,
     legend="insidetop",meltvar="variable",flip=FALSE,ptsize=3)
}
\arguments{
\item{indt}{Input data.frame with at least one date variable and one or more vategorical variables, if melted.}

\item{title, xlab, ylab}{TItles and Labels}

\item{breaks}{A list or text as follows
\itemize{
\item \verb{<doiset>} : A dates of interest category, see \code{\link[=fg_get_dates_of_interest]{fg_get_dates_of_interest()}}
\item list of integers: A list of days for which to go back in time, e.g. \code{c(7,30,360)} creates intervals for the last week, 1 week to 1 month, etc.
\item list of reals in \verb{[0,1]} fractions of the dates in each category, e.g \code{c(0.2,1)} creates intevals with the last 20pct of dates, and any older.
}}

\item{doi}{Points or segments to overlay with latest observations, or changes since a particular date.
\itemize{
\item \code{"last"} (Default)  Last date as a dot
\item \code{"last,<d1>"} Segment from date \code{d1} to last date in input data.
\item \code{"last,n"} Segment from \code{n}th date from the end to the end.
\item \code{"date,<d1>"} Levels as of date \code{d1}
\item \code{"none"}  No points or segments.
}}

\item{normalize}{Normalize data in some way prior to plotting.  Choices are
\itemize{
\item \code{"byhistcat"}Transform data into percentiles within each variable and historical category
\item \verb{"byvar","zbyvar"}Transform data into percentiles (or z-scores) within each variable and historical category
}}

\item{orderby}{(Default \code{""}) Underlying categories are by default ordered as in \code{indt}, unless
\itemize{
\item \verb{"value","-value"} : Order by last value in series for each category or descending if \code{"-value"}
\item \verb{"date,<d1>","-date,<d1>"} : Order by value (or decreasing value) at date \verb{<d1>}
\item \verb{"alpha","-alpha"} Order alphabetically in ascending or descnding orer.
}}

\item{boxtype}{Formatting of boxplots.  If in \verb{"violin},\code{"viobycat"} make a violin plot, otherwise show a full boxplot (with outliers turned off by default), with any aspects in \code{c("nostaple","nomedian","nobox")} taken out.}

\item{dropset}{String or list with underlying categories to drop from graph}

\item{hilightcats}{String or list of underlying categories to highlight with differnt color in label.}

\item{addline}{in \code{c("mean","last")} Add a horizontal line across the mean of all observations or a smooth like across last observations}

\item{facetform}{(Default: "") Any faceting formula which includes text or factor columns in \code{indt}. See examples and note that facets can also be added using \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} to the output graph.}

\item{ycoord}{(Default NULL) a two element list with limits on y corrdinates}

\item{trimpctile}{(Default 0) trims data before any plotting to fall within \code{c(trimpctile,1-trimpctile)} percentiles within each variable.}

\item{legend}{(Default \code{"insidetop"}) Where to put the legend}

\item{meltvar}{(Default: \code{"variable"} Name of variable with unit category.}

\item{flip}{(Default \code{FALSE}) If \code{TRUE} then categories are arranged vertically}

\item{ptsize}{(Default: 3) Size of points for \code{doi} parmeter}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plots static summaries of time series in boxplot form.
}
