% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_graphs.R
\name{fgts_dygraph}
\alias{fgts_dygraph}
\title{TIme series in Dygraph form}
\usage{
fgts_dygraph(
  indt,
  title = "",
  ylab = "",
  roller = "default",
  pointers = "hair,both",
  splitcols = FALSE,
  stepcols = FALSE,
  hidecols = FALSE,
  hilightcols = FALSE,
  hilightwidth = 2,
  hilightstyle = "solid",
  events = "",
  event_ds = NULL,
  annotations = "",
  annotation_ds = NULL,
  forecast_ds = NULL,
  ylimits = NULL,
  dtstartfrac = 0,
  dtwindow = "",
  exportevents = NULL,
  meltvar = "variable",
  dylegend = "always",
  groupnm = "common",
  fillGraph = FALSE,
  verbose = FALSE,
  extraoptions = list()
)
}
\arguments{
\item{indt}{Input data in long or wide format.  THere must be at least one date column, one
character column and one numeric column.  Ideal format is \verb{date,variable,value}}

\item{title}{Title to put on top of graph}

\item{ylab}{Label for y axis}

\item{roller}{Initial moving average value to smooth graphs.  (See \code{\link[dygraphs:dyRoller]{dygraphs::dyRoller()}})  Options are
\itemize{
\item \code{default} (Default) chose a smoothing parameter consistent with the length of the input series
\item \code{finest} No smoothing
\item integer >= 0 : User specified moving average length.
}}

\item{pointers}{Pointer options to use with mouse movement. (See \code{\link[dygraphs:dyCrosshair]{dygraphs::dyCrosshair()}})}

\item{splitcols, stepcols, hidecols}{String or list of data series to show on a second y axis, to be shown as step plots, or to be hidden.}

\item{hilightcols}{String or list of data series to plot in different style than other series.}

\item{hilightwidth}{(Default: 2) relative width of series specified in \code{hilightcols}}

\item{hilightstyle}{(Default: solid).  Line style of series specified in \code{hilightcols}.
Options are (\code{solid},\code{dashed},\code{dotted},\code{dotdash})}

\item{events}{String with possible events to add to graph.  Options can be added together
with \verb{;} and include
\itemize{
\item \verb{doi,eventsetname}  : Events in internal event list \code{eventsetname} from list maintained by \code{\link[=fg_update_dates_of_interest]{fg_update_dates_of_interest()}}.
\item \verb{seasonal,type} : Regularly spaced intervals of \code{type}.  For options, see details.
\item \code{minmax} : Locations of highest and lowest observations per series.
\item \verb{dt,text,d1,<d2>} : Text events starting at \code{d1} and possibly ending at \verb{<d2>},both
of the form \code{yyyy-mm-dd}.
\item \verb{break,labelform} : Breakouts as determined by \code{\link[=fg_addbreakouts]{fg_addbreakouts()}} with \code{labelform} in ("singleasdate","singleavalue","breakno")
\item \verb{tp,n} : Turning points as determined by \code{\link[=fg_findTurningPoints]{fg_findTurningPoints()}}
}}

\item{event_ds}{\code{data.frame} of events to be added to graph.  See details and
examples for specification.}

\item{annotations}{string with annotations on individual series or along \code{y} axes.  Options can be added
together with \verb{;}  and can include
\itemize{
\item \code{lastvalue}  : Value of latest observation for each series, placed at the end of the series
\item \code{lastlabel}  : Name of each series, placed at the end of the series.
\item \verb{hline,y : Horizontal line at }y`
\item \verb{range,ybeg,yend} : Band placed between \code{ybeg} and \code{yend}
}}

\item{annotation_ds}{\code{data.frame} of annotations added to graph. See details for specification.}

\item{forecast_ds}{\code{data.frame} of forecasts to be displayed after the end of those in \code{indt}. Those typically are in wide format, with at minimum
a (first) date column and series names of the form \verb{series.f``, }series.flo\verb{ and/or `series.fhi}, where \code{series} is one of the
plotted series in \code{indt}}

\item{ylimits}{Two number vector of lower and upper limits of data to be displayed. Alternatively,
a string of the form \verb{<seriesnm>,<q>} will limit displayed data to the (q,1-q) quantiles of \code{seriesnm}}

\item{dtstartfrac}{Fraction in (0,1] of dates in \code{indt} to start the range selector.
See \code{\link[dygraphs:dyRangeSelector]{dygraphs::dyRangeSelector()}}}

\item{dtwindow}{String to specify date ranges applied \code{\link[dygraphs:dyRangeSelector]{dygraphs::dyRangeSelector()}} of the
form \code{begin::end} where either end can take the form "yyyy-mm-dd" or a relative date to the other end of the
series, e.g \verb{-3m} or \verb{-2w}.  Example: \code{"-3m::-1m"} defines a 2 month period 1 month back from the end of the series.}

\item{exportevents}{String of name of \code{data.frame} to create in  \code{.GlobalEnv} with event dates displayed on graph.}

\item{meltvar}{(Default: \code{variable}) Column name in \code{indt} with series names, if melted.}

\item{dylegend}{(Default: TRUE) include legend in graph}

\item{groupnm}{(Default: \code{common})  Group name used in \code{shiny} or \code{RMarkdown} to synchonize graphs}

\item{fillGraph}{(Default: FALSE) Shade area underneath each series.}

\item{verbose}{(Default: FALSE) Print extra details about what will be graphed.}

\item{extraoptions}{Additional options passed to \code{\link[dygraphs:dyOptions]{dygraphs::dyOptions()}}}
}
\value{
Dygraph of data
}
\description{
TIme series in Dygraph form
}
\details{
Input data can either be in wide ('date' ,'series1',...) format or normalized (long) format
('date','variable','value') format.  This package infers date columns names from column types and seeks to be as agnostic
as possible as to column names.
Colors can be managed using \code{\link[=fg_update_colors]{fg_update_colors()}} and will persist across R sessions,
Series are grouped together into bands around a series \code{series} if their names end as in 'series.lo' or 'series.hi'.  See examples and vignette for details.

\strong{Events} are dates and date ranges to be highlighted in the graph.   Days of interest (doi) can be added
using \code{\link[=fg_update_dates_of_interest]{fg_update_dates_of_interest()}} which will persist across R sessions.  See examples and vignette for further details.
Events can also be added using a \code{data.frame} passed via \code{event_ds} with the following columns:

\preformatted{  xx zz 111}\tabular{lccl}{
   column \tab required \tab type \tab description \cr
   \code{date} \tab yes \tab Date \tab Start date \cr
   \code{date_end} \tab  \tab Date \tab End date to specify range of a colored band \cr
   \code{text} \tab yes \tab character \tab Text to display \cr
   \code{color} \tab  \tab  \tab character \cr
   \code{eventonly} \tab  \tab logical \tab Only draw line for for start of event, no band \cr
   \code{strokePattern} \tab  \tab character \tab \code{dashed} (default), one of ('solid','dashed','dotted','dotdash') \cr
   \code{loc} \tab  \tab character \tab 'bottom' (default), one of ('top','bottom') \cr
   \code{series} \tab  \tab character \tab Name of series to apply event to, if needed \cr
}


Many times, events depend on outside data or statistical analysis on the original data.  The \code{event_ds} to be passed
in can come from event helpers in \code{\link[=overlay_eventset]{overlay_eventset()}}, \code{\link[=fg_addbreakouts]{fg_addbreakouts()}}, \code{\link[=fg_findTurningPoints]{fg_findTurningPoints()}}, or  \code{\link[=fg_ratingsEvents]{fg_ratingsEvents()}}.

\strong{Annotiations} include any notes or highlights added to the graph on the 'y' axis or on an individual series.  In addition to those passed
via the \code{annotations} parameter, annotations can be added using a \code{data.frame} with the following columns:\tabular{lccl}{
   column \tab required \tab type \tab description \cr
   \code{date} \tab yes \tab Date \tab Start date \cr
   \code{date_end} \tab  \tab Date \tab End date to specify range of a colored band \cr
   \code{text} \tab yes \tab character \tab Text to display \cr
   \code{color} \tab  \tab  \tab character \cr
   \code{eventonly} \tab  \tab logical \tab Only draw line for for start of event, no band \cr
}
}
\examples{
fgts_dygraph(eqtypx[,.(date,TLT)], title="Stock Prices", events="doi,regm;doi,fedmoves")

}
