% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_graphs.R
\name{fgts_dygraph}
\alias{fgts_dygraph}
\title{TIme series in Dygraph form}
\usage{
fgts_dygraph( indt,
 title = "",  xlab="", ylab = "",  roller = "default",  bg_opts = "hair,both;grid,both",
 splitcols = FALSE, stepcols = FALSE, hidecols = FALSE, hilightcols = FALSE,
 hilightwidth = 2, hilightstyle = "solid",
 events = "", event_ds = NULL,
 annotations = "", annotation_ds = NULL, forecast_ds = NULL,
 ylimits = NULL,  dtstartfrac = 0,  dtwindow = "",  rebase = "",  exportevents = NULL,
 meltvar = "variable",  dylegend = "always",  groupnm = fg_sync_group(),  fillGraph = FALSE,
 verbose = FALSE,  extraoptions = list() )
}
\arguments{
\item{indt}{Input data in long or wide format.  THere must be at least one date column, one
character column and one numeric column.  Ideal format is \verb{date,variable,value}}

\item{title}{Title to put on top of graph}

\item{xlab, ylab}{Labels for x and y axis}

\item{roller}{Initial moving average value to smooth graphs.  (See \code{\link[dygraphs:dyRoller]{dygraphs::dyRoller()}})  Options are
\itemize{
\item \code{default} (Default) chose a smoothing parameter consistent with the length of the input series
\item \code{finest} No smoothing
\item integer >= 0 : User specified moving average length.
}}

\item{bg_opts}{Semicomma separated options to change interactivity and background of charts.  These semicomma separated options
change pointer option and grids options.
\itemize{
\item \verb{hair,<style>} passes \verb{<style>} to \code{\link[dygraphs:dyCrosshair]{dygraphs::dyCrosshair()}}.  Default is "both" x and y crosshairs.
\item \verb{grid,<x,y,both>} specifies which gri lines to show.  Default is "both"
\item \code{norange} turns off date range selector.
}}

\item{splitcols, stepcols, hidecols}{String or list of data series to show on a second y axis, to be shown as step plots, or to be hidden.
Can also be \code{TRUE} in which case first series in the data is affected. Can also be a semicolon separated single string with mutiple series.}

\item{hilightcols}{String or list of data series to plot in different style than other series.}

\item{hilightwidth}{(Default: 2) relative width of series specified in \code{hilightcols}}

\item{hilightstyle}{(Default: solid).  Line style of series specified in \code{hilightcols}.
Options are (\code{solid},\code{dashed},\code{dotted},\code{dotdash})}

\item{events}{String with possible events to add to graph.  Options can be added together
with \verb{;} and include
\itemize{
\item \verb{doi,<eventsetname>}  : Events in internal event list \code{eventsetname} from list maintained by \code{\link[=fg_update_dates_of_interest]{fg_update_dates_of_interest()}}.
\item \verb{seasonal,<type>} : Regularly spaced intervals of dates. See details below.
\item \code{minmax} : Locations of highest and lowest observations per series.
\item \verb{dt,text,d1,<d2>} : Text events starting at \code{d1} and possibly ending at \verb{<d2>},both
of the form \code{yyyy-mm-dd}.
\item \verb{break,labelform} : Breakouts as determined by \code{\link[=fg_addbreakouts]{fg_addbreakouts()}} with \code{labelform} in ("singleasdate","singleavalue","breakno")
\item \verb{tp,n} : Turning points on the first series as determined by \code{\link[=fg_findTurningPoints]{fg_findTurningPoints()}}
}}

\item{event_ds}{\code{data.frame} of events to be added to graph.  See details and
examples for specification.}

\item{annotations}{string with annotations on individual series or along \code{y} axes.  Options can be added
together with \verb{;}  and can include
\itemize{
\item \code{lastvalue}  : Value of latest observation for each series, placed at the end of the series
\item \code{lastlabel}  : Name of each series, placed at the end of the series.
\item \verb{hline,y : Horizontal line at }y`
\item \verb{range,ybeg,yend} : Band placed between \code{ybeg} and \code{yend}
}}

\item{annotation_ds}{\code{data.frame} of annotations added to graph. See details for specification.}

\item{forecast_ds}{\code{data.frame} of forecasts to be displayed after the end of those in \code{indt}. Those typically are in wide format, with at minimum
a (first) date column and series names of the form \verb{series.f``, }series.flo\verb{ and/or `series.fhi}, where \code{series} is one of the
plotted series in \code{indt}}

\item{ylimits}{Two number vector of lower and upper limits of data to be displayed. Alternatively,
a string of the form \verb{<seriesnm>,<q>} will limit displayed data to the (q,1-q) quantiles of \code{seriesnm}}

\item{dtstartfrac}{Fraction in (0,1] of dates in \code{indt} to start the range selector.
See \code{\link[dygraphs:dyRangeSelector]{dygraphs::dyRangeSelector()}}}

\item{dtwindow}{String to specify date ranges applied \code{\link[dygraphs:dyRangeSelector]{dygraphs::dyRangeSelector()}} of the
form \code{begin::end} where either end can take the form "yyyy-mm-dd" or a relative date to the other end of the
series, e.g \verb{-3m} or \verb{-2w}.  Example: \code{"-3m::-1m"} defines a 2 month period 1 month back from the end of the series.}

\item{rebase}{String of the form \verb{yyyy-mm-dd,<value>} with \verb{<value>} assumed 100 if not specified.  This normalizes all series to \verb{<value>}
as of the given date.  See examples.}

\item{exportevents}{String of name of \code{data.frame} to create in  \code{.GlobalEnv} with event dates displayed on graph.}

\item{meltvar}{(Default: \code{variable}) Column name in \code{indt} with series names, if melted.}

\item{dylegend}{(Default: "auto") Passed to \code{\link[dygraphs:dyLegend]{dygraphs::dyLegend()}}, can be one of ("auto", "always", "onmouseover", "follow", "never")}

\item{groupnm}{(Default: NULL, unless set via \code{\link[=fg_sync_group]{fg_sync_group()}} )  Group name used in \code{shiny} or \code{RMarkdown} to synchronize graphs. See \code{\link[=fg_sync_group]{fg_sync_group()}} for details.}

\item{fillGraph}{(Default: FALSE) Shade area underneath each series.}

\item{verbose}{(Default: FALSE) Print extra details about what will be graphed.}

\item{extraoptions}{Additional options passed to \code{\link[dygraphs:dyOptions]{dygraphs::dyOptions()}}}
}
\value{
Dygraph \href{https://rstudio.github.io/dygraphs/}{dygraphs} of input data, with annotations and other customizations.
}
\description{
TIme series in Dygraph form
}
\details{
Input data can either be in wide ('date' ,'series1',...) format or normalized (long) format
('date','variable','value') format.  This package infers date columns names from column types and seeks to be as agnostic
as possible as to column names.
Colors can be managed using \code{\link[=fg_update_colors]{fg_update_colors()}} and will persist across R sessions, See vignette for details.
Series are grouped together into bands around a series \code{series} if their names end as in 'series.lo' or 'series.hi'.  See examples and vignette for details.

\strong{Events} are dates and date ranges to be highlighted in the graph. Multiple types of events can be strung together in
semicolon delimited strings. Of the options outlined above, two additional details are
\itemize{
\item \code{"doi,<category>"} gets events from \code{\link[=fg_get_dates_of_interest]{fg_get_dates_of_interest()}} which can be added to or managed using \code{\link[=fg_update_dates_of_interest]{fg_update_dates_of_interest()}}.
Colors and label placement can be customized as necessary.
\item \verb{seasonal,<type>} puts regularly occurring events on the graph.  \verb{<type>} can be
}\tabular{ll}{
   \verb{<type>} \tab description \cr
   \verb{"optex,mo} | \verb{qtr"} \tab Monthly and/or quarterly equity option expiration dates. \cr
   \code{"roll"} \tab IMM CDS roll dates \cr
   \code{"daysfromroll"} \tab Dates with same number of days to the next roll as last date plotted \cr
   \verb{"doq","doy","bdoy"} \tab Dates with same day in quarter, in year, or business day of year to the last day plotted \cr
}


Events can also be added using a \code{data.frame} passed via \code{event_ds} with the following columns:\tabular{llc}{
   column \tab description \tab type \cr
   \code{date} \tab (Required) Start date \tab \code{Date} \cr
   \code{date_end} \tab End date to specify range of a colored band \tab \code{Date} \cr
   \code{text} \tab (Required) Text to display \tab \code{character} \cr
   \code{color} \tab Color for line and text \tab \code{character} \cr
   \code{eventonly} \tab Only draw line for for start of event, no band \tab \code{logical} \cr
   \code{strokePattern} \tab One of ('solid','dashed' (Default) ,'dotted','dotdash') \tab \code{character} \cr
   \code{loc} \tab one of ('top','bottom' (Default)) \tab \code{character} \cr
   \code{series} \tab Name of series to apply event to, if needed \tab \code{character} \cr
   \code{category} \tab Optional string used for exceptions. See notes below. \tab \code{character} \cr
}


Many times, events depend on outside data or statistical analysis on the original data.  The \code{event_ds} to be passed
in can come from event helpers in \code{\link[=fg_cut_to_events]{fg_cut_to_events()}}, \code{\link[=fg_addbreakouts]{fg_addbreakouts()}}, \code{\link[=fg_findTurningPoints]{fg_findTurningPoints()}}, or  \code{\link[=fg_ratingsEvents]{fg_ratingsEvents()}}.
Event columns are processed as is, unless \code{category=="series_color"} which will replace \code{color} with that of its series.

\strong{Annotiations} include any notes or highlights added to the graph on the 'y' axis or on an individual series.  In addition to those passed
via the \code{annotations} parameter, annotations can be added using a \code{data.frame} with the following columns:\tabular{llc}{
   column \tab description \tab type \cr
   \code{date} \tab (Required) Start date \tab \code{Date} \cr
   \code{date_end} \tab End date to specify range of a colored band \tab \code{Date} \cr
   \code{text} \tab (Required) Text to display \tab \code{character} \cr
   \code{color} \tab Color for line and text \tab \code{character} \cr
   \code{eventonly} \tab Only draw line for for start of event, no band \tab \code{logical} \cr
}


Other notes:
\itemize{
\item Using \code{stepcols} most often happens with lower frequency data, so an \code{nafill} is automatically performed.
}
}
\examples{
# See Vignette for more extensive examples.
# Basic Example
fgts_dygraph(eqtypx, title="Stock Prices", ylab="Adjusted Close")

# With series Highlights, finer resolution and focused date range
fgts_dygraph(eqtypx, dtstartfrac=0.8,hilightcols="IBM",hilightwidth=4,roller=3)

# Rebasing to 1/1/2022
fgts_dygraph(eqtypx, title="Rebased Prices", ylab="Adjusted Close",rebase="2022-01-01")

# Using bands (.lo, .hi)
toplot <- reerdta[REGION=="LATAM",.(cop=sum(value*(variable=="COL")),
              reer=mean(value),reer.lo=min(value),reer.hi=max(value)),by=.(date)]
fgts_dygraph(toplot,title="COP REER vs Latam peers",roller=3,hilightcols="cop",hilightwidth=4)

# Events Examples.  Notice how roller shortens with the series.
# See Vignette for more extensive examples
require(data.table)
smalldta <- narrowbydtstr(eqtypx[,.(date,TLT,EEM)],"-3y::")
fgts_dygraph(smalldta,events="doi,regm;doi,fedmoves")
fgts_dygraph(smalldta,events="date,FOMO,2025-01-01,2025-06-01;date,xmas,2025-12-25")

# Events passed in as data.frames
myevents = data.frame(end_date=as.Date(c("2024-03-10","2024-06-10")),
             date=as.Date(c("2024-01-10","2024-04-10")),
             text=c("range","event"),color=c("green","red"))
fgts_dygraph(smalldta,events="doi,fedmoves",event_ds=myevents)

# Annotations on y axis
fgts_dygraph(eqtypx,annotations="last,linevalue")
fgts_dygraph(eqtypx,annotations="hline,100,at100,red;hline,200,at200;range,300,400")

# use with helpers
require(data.table)
smalldta <- narrowbydtstr(eqtypx[,.(date,IBM,QQQ)],"-2y::")
fgts_dygraph(smalldta,title="W TurnPts",event_ds=fg_findTurningPoints(smalldta[,.(date,QQQ)]))
fgts_dygraph(smalldta,title="W Sentiment",event_ds=fg_cut_to_events(consumer_sent,center="zscore"))
fgts_dygraph(smalldta,title="W dividends",event_ds=fg_tq_divs(c("IBM","QQQ")))

# Other helpers for use with credit ratings, breakouts, and earnings data are available.

# use with forecasts

require(forecast)
smalldta <- narrowbydtstr(eqtypx[,.(date,IBM,QQQ)],"-2y::")
fcst_one <- function(ticker) {
  t1_ts <- zoo::zoo(smalldta[[ticker]],smalldta[["date"]])
  forecast::ets(t1_ts) |> forecast::forecast(h=36) |>  fg_predict(seriesnm=ticker)
  }
fpred <- merge(fcst_one("QQQ"),fcst_one("IBM"),by="date")
fgts_dygraph(smalldta,title="With Forecasts", dtstartfrac=0.7,forecast_ds=fpred)

}
