% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinanceGraphs-package.R
\docType{package}
\name{FinanceGraphs-package}
\alias{FinanceGraphs}
\alias{FinanceGraphs-package}
\title{FinanceGraphs: Flexible Graphs for Analysis of Financial Data and Time Series}
\description{
Flexible wrappers around R graphics modules [dygraphs] and [ggplot2] to visualize time-based data. Interactive time series plots include multiple options for including external data such as forecasts and events. Other static plots useful for time series data include boxplots for multiple categories, event study graphs for analysis around a set of dates, and scatter plots to visualize bivariate time series data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/derekholmes0/FinanceGraphs}
  \item Report bugs at \url{https://github.com/derekholmes0/FinanceGraphs/issues}
}

}
\author{
\strong{Maintainer}: Derek Holmes \email{derek@derekholmes.com} [copyright holder]

}
\keyword{internal}
