% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinanceGraphs-package.R
\docType{package}
\name{FinanceGraphs-package}
\alias{FinanceGraphs}
\alias{FinanceGraphs-package}
\title{FinanceGraphs: Flexible Graphs for Analysis of Financial Data and Time Series}
\description{
Flexible wrapper and helper functions for display of Time Series using [dygraphs]. Helpers for adding date-specific events and ranges are included, as well as for adding forecasts and other annotations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/derekholmes0/FinanceGraphs}
  \item Report bugs at \url{https://github.com/derekholmes0/FinanceGraphs/issues}
}

}
\author{
\strong{Maintainer}: Derek Holmes \email{derek@derekholmes.com} [copyright holder]

}
\keyword{internal}
