% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinanceGraphs-package.R
\docType{package}
\name{FinanceGraphs-package}
\alias{FinanceGraphs}
\alias{FinanceGraphs-package}
\title{FinanceGraphs: Flexible Graphs for Analysis of Financial Data and Time Series}
\description{
Flexible graphs for analysis of Financial Data and Time Series
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/derekholmes0/FinanceGraphs}
  \item Report bugs at \url{https://github.com/derekholmes0/FinanceGraphs/issues}
}

}
\author{
\strong{Maintainer}: Derek Holmes \email{derek@derekholmes.com} [copyright holder]

}
\keyword{internal}
