% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constants.R
\name{fg_get_aes}
\alias{fg_get_aes}
\alias{fg_get_aesstring}
\alias{fg_display_colors}
\alias{fg_print_aes_list}
\alias{fg_get_datemap}
\alias{fg_current_theme}
\title{Maintain Colors}
\usage{
fg_get_aes(item = "", n_max = NA_integer_, asdataframe = FALSE)

fg_get_aesstring(
  item = "",
  n_max = NA_integer_,
  toget = "value",
  rtnifnotfound = FALSE
)

fg_display_colors(item = "")

fg_print_aes_list(grepstr = "")

fg_get_datemap()

fg_current_theme()
}
\arguments{
\item{item}{(Default: "") A grep string for categories desired.}

\item{n_max}{Maximum number of rows or entries to return.  Required for \code{Rcolorbrewer} color aesthetics}

\item{asdataframe}{(Default: FALSE) Return dataframe of parameters regardless of type. (See details)}

\item{toget}{Column in the aes \code{data.frame} to paste together as a string.}

\item{rtnifnotfound}{Return \code{NA_character_} if aes not found}

\item{grepstr}{narrow list of internal aesthetics sets to functions from \code{grepstr}}
}
\value{
\code{fg_get_aes()} returns \code{data.frame} of aesthetics, including sorting columns, help strings, and values,
\code{fg_get_aesstring()} returns a list with just the character values of the requested aesthetic.
\code{fg_print_aes_list()} returns a markdown ready character vector of aesthetic names used in each function
\code{fg_display_colors()} returns a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object with colors and associated names for an aesthetic name
}
\description{
\code{fg_get_aes()} gets aethestic  \code{data.frame} for use in graphs.
\code{fg_get_aesstring()} takes a column from the \code{data.frame} retrieved by \code{fg_get_aes()}
\code{fg_print_aes_list()} prints names of aesthetics used internally in FinanceGraph functions.
\code{fg_display_colors()} Shows a plot with current colors.
\code{fg_get_datemap()} returns a data.table of all dates and date mappings
\code{fg_current_theme()} returns the current \code{ggplot2} theme used for scatter plots.
}
\examples{
# Data set, String
head(fg_get_aes("lines"),3)
fg_get_aesstring("lines")
#  Gradient colors are stored in a `data.frame` as in a set of "Blue Greens"
fg_get_aes("espath_gp",asdataframe=TRUE)
# To get the actual colors, we need to know how many:
fg_get_aes("espath_gp",n_max=8)
fg_display_colors("lines")

}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}, \code{\link[=fg_update_aes]{fg_update_aes()}}
}
