% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{reerdta}
\alias{reerdta}
\title{Nominal FX levels}
\format{
\subsection{\code{reerdta}}{

A tibble with 1920 observations and 3 columns
\describe{
\item{date}{Date Of Observation}
\item{variable}{ISO Country code}
\item{REGION}{investment region to which each country belongs}
\item{value}{Index of Real Effective Exchange Rates}
}
}
}
\source{
\url{https://imf.org}
}
\usage{
reerdta
}
\description{
Real Effective Exchange Rates
}
\keyword{datasets}
