% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_constants.R
\name{fg_update_aes}
\alias{fg_update_aes}
\alias{fg_update_line_colors}
\alias{fg_replace_theme}
\alias{fg_verbose}
\alias{fg_reset_to_default_state}
\title{Maintain Colors}
\usage{
fg_update_aes(indta, aestype = NA_character_, persist = TRUE, replace = FALSE)

fg_update_line_colors(colorlist, replace = FALSE, persist = TRUE)

fg_replace_theme(newTheme, persist = TRUE)

fg_verbose(item = "")

fg_reset_to_default_state(reset = "all")
}
\arguments{
\item{indta}{\code{data.table} aesthetic \code{data.fram} with columns as shown in details.}

\item{aestype}{(Default: \code{NA}) character string with type of aesthetic requested.  If not provided in \verb{[fg_Update_aes()]} the}

\item{persist}{(Default: TRUE) Keep changes across invocations of the package.}

\item{replace}{(Default: FALSE) Replaces existing dates of interest with new set provided, otherwise replaces/inserts new rows only.}

\item{colorlist}{List with up to 14 new colors just for line (series) coloring}

\item{newTheme}{A new ggplot2 theme}

\item{item}{(Default: "") A grep string for categories desired.}

\item{reset}{(Default: "all"), options in ("all","colors","doi") to reset to defaults with the package.}
}
\value{
No return value, as these are called for the side effects of adding to or replacing aesthetic sets.
}
\description{
\code{fg_update_aes()} updates or replaces default aesthestics (e.g. colors, linestyles, etc).
\code{fg_update_line_colors()} replaces line colors only
\code{fg_reset_to_default_state()} resets colors and/or dates of interest
\code{fg_replace_theme()} Replaces default theme used in static plots
\code{fg_verbose()} Toggles printing of aesthetics
}
\details{
For colors,
New data passed into \code{\link[=fg_update_aes]{fg_update_aes()}} persists across future loads of the package unless \code{persist=FALSE}.
New color datasets must have at least three columns:\tabular{ll}{
   Column \tab Meaning \cr
   \code{category} \tab Arbitrary aestehtic category, e.g. \code{"lines"} for line colors. \cr
   \code{variable} \tab Any string that can be sorted or grepped to map to data. \cr
   \code{type} \tab Aesthetic type, in \code{c("color","colorrange","linetype","symbol","alpha")} \cr
   \code{value} \tab String with value detired (e.g a color) \cr
}


\code{variable} is used to prioritize colors, so (e.g. \code{D01} will be the color of the first series in an input dataset)

If \code{aestype=="colorrange"} then a sequential scale of size \code{n_max} will be returned using details saved from  \code{\link[=fg_update_aes]{fg_update_aes()}}. See \link[scales:pal_brewer]{scales::brewer_pal}
and \href{https://colorbrewer2.org/#type=sequential&scheme=Greens&n=7}{colorbrewer}
}
\examples{
# Data set, String
head(oldcolors <- fg_get_aes("lines"),3)
# then change as needed.  For example, to make the second line blue, and the 4th line red,
oldcolors[c(2,3),"value"] <- c("blue","tomato")
fg_update_aes( oldcolors )
head( fg_get_aes("lines"),3)
# to create a new category, make a similar `data.frame`, as in
newcolors <- data.frame(category=rep("mylines",3),variable=c("D01","D02","D03"),
                value=c("red","black","green"))
fg_update_aes( newcolors, aestype="color")
fg_get_aesstring("mylines")
THeme replacement
require(ggplot2)
fg_replace_theme(ggplot2::theme_dark(),persist=FALSE)
fg_reset_to_default_state()

}
\seealso{
\code{\link[=fgts_dygraph]{fgts_dygraph()}}, \code{\link[=fg_scatplot]{fg_scatplot()}}, \code{\link[=fg_get_aes]{fg_get_aes()}}
}
