% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yc_CMSUST}
\alias{yc_CMSUST}
\title{Constant Maturity UST rates}
\format{
\subsection{\code{yc_CMSUST}}{

A long format \code{data.table} with constant maturity UST with tenors 2 year, 10 year and 30 year
\describe{
\item{variable}{Term of UST}
\item{date}{Date of observation}
\item{value}{Annualized Percent interest rae}
}
}
}
\source{
\url{https://fred.stlouisfed.org/}
}
\usage{
yc_CMSUST
}
\description{
FRED calculated constant maturity interest rates
}
\keyword{datasets}
